Added config checks before operation.

+++ Dependencies/StorageCore/StorageCore/CoreDataStore.swift
@@ -8,11 +8,12 @@
 import Foundation
 import CoreData
 
-open class CoreDataStore: CoreDataConfigProvider & CoreDataOperationProvider {
+open class CoreDataStore: CoreDataConfigProvider {
     
     private(set) public var managedContext: NSManagedObjectContext!
     private(set) public var persistentContainer: NSPersistentContainer!
-    let containerName: String
+    private(set) public var containerName: String
+    private(set) public var isConfigured: Bool = false
     
     public init(containerName: String) {
         self.containerName = containerName
@@ -21,6 +22,7 @@ open class CoreDataStore: CoreDataConfigProvider & CoreDataOperationProvider {
     open func configureSetups(onResponse: @escaping (Error?) -> Void) {
         self.loadPersistentStores(with: containerName) { (storeDescription, error) in
             self.managedContext = self.persistentContainer.viewContext
+            self.isConfigured = (error == nil)
             
             onResponse(error)
         }
@@ -30,21 +32,34 @@ open class CoreDataStore: CoreDataConfigProvider & CoreDataOperationProvider {
                                      onResponse: @escaping (NSPersistentStoreDescription, Error?) -> Void) {
         let container = NSPersistentContainer(name: persistentContainerName)
         container.loadPersistentStores { (storeDescription, error) in
+            self.persistentContainer = container
             
-            if error == nil {
-                self.persistentContainer = container
-            }
-            
+            onResponse(storeDescription, error)
         }
     }
     
     public func saveManagedContext(onResponse: (Bool, Error?) -> Void) {
-        
+        if self.managedContext.hasChanges {
+            do {
+                try self.managedContext.save()
+                onResponse(true, nil)
+            } catch {
+                onResponse(false, error)
+            }
+        }
     }
+    
+}
+
+extension CoreDataStore: CoreDataOperationProvider {
     
     public func fetch<MANAGED_OBJECT: NSManagedObject>(_ managedObjectType: MANAGED_OBJECT.Type,
                                                        filter predicate: NSPredicate? = nil,
                                                        onResponse: (_ object: [MANAGED_OBJECT]?, _ error: Error?) -> Void) {
+        if !self.isConfigured {
+            onResponse(nil, CoreDataStoreError.notYetConfigured)
+            return
+        }
         
         let objectEntityName: String = String(describing: managedObjectType)
         let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: objectEntityName)
@@ -61,6 +76,11 @@ open class CoreDataStore: CoreDataConfigProvider & CoreDataOperationProvider {
     
     public func delete(_ managedObject: NSManagedObject,
                 onResponse: (Bool, Error?) -> Void) {
+        if !self.isConfigured {
+            onResponse(false, CoreDataStoreError.notYetConfigured)
+            return
+        }
+        
         self.managedContext.delete(managedObject)
         self.saveManagedContext(onResponse: onResponse)
     }
